+++ restinparse-core/src/main/java/com/github/jmfayard/model/QueryResults.java
@@ -6,12 +6,16 @@ import java.util.List;
 
 public class QueryResults {
     public final List<ParseMap> results;
+    public final Long count;
 
-    public QueryResults(List<ParseMap> results) {
+    public QueryResults(List<ParseMap> results, Long count) {
+        this.count = count;
         if (results == null) {
             this.results = Collections.emptyList();
         } else {
             this.results = new ArrayList<ParseMap>(results);
         }
     }
+
+
 }